add main in AM
